+++ term/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "term"
+version = "0.1.0"
+edition = "2021"
+publish = false
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
